src/ITMDashboard/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==1.1.2
idna==3.3
influxdb==5.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
msgpack==0.6.1
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.3
